+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2020-08-08
+### Changed
+- **Breaking:** `SearchQueryBloc` is now `SearchQueryCubit`
+  - The `SetSearchQuery` event is now the `setQuery` method
+  - The `ClearSearchQuery` event is now the `clearQuery` method
+
 ## [0.2.0] - 2020-07-28
 ### Added
 - Changelog documentation